Updated makefile for debug builds and osSyncPrintf Removed memory.h since memcpy is in libc string.h Removed graphics.c Removed graphics_context.c Removed some debug messages Removed files from ./include Aligned graphics_context_t to 16 bytes and the framebuffers to 64 bytes Set video mode to Hires, Point Filtered Filtered deinterlacing, 16bit color Changed printf to osSyncPrintf

+++ axis/main.h
@@ -4,11 +4,11 @@
 #include "video.h"
 #include "stdint.h"
 
-#define	STATIC_SEGMENT	1
-#define	RSP_FB_SEGMENT	2
+#define	STATIC_SEGMENT 1
+#define	RSP_FB_SEGMENT 2
 
-#define	STACKSIZE	0x2000
-#define STACKSIZE_NUM (STACKSIZE / sizeof(unsigned long long))
+#define	STACKSIZE 0x2000
+#define STACKSIZE_NUM (STACKSIZE / sizeof(uint64_t))
 
 #ifdef _LANGUAGE_C
